+++ tests/src/Functional/AllProfileInstallationTestsAbstract.php
@@ -26,6 +26,13 @@ use Drupal\Tests\BrowserTestBase;
 abstract class AllProfileInstallationTestsAbstract extends BrowserTestBase {
 
   /**
+   * Whether to strictly check schema on installed configuration.
+   *
+   * @var bool
+   */
+  protected $strictConfigSchema = FALSE;
+
+  /**
    * Override the default drupalLogin method.
    *
    * The openid_connect modules is configured to alter the user.login route and
@@ -97,6 +104,7 @@ abstract class AllProfileInstallationTestsAbstract extends BrowserTestBase {
     $this->ensureWebformInstall();
     $this->ensurePublishContentInstalled();
     $this->ensureEventFeatureInstalled();
+    $this->ensurePromotionsFeatureInstalled();
   }
 
   /**
@@ -275,6 +283,22 @@ abstract class AllProfileInstallationTestsAbstract extends BrowserTestBase {
 
     // Ensure the taxonomy is accessible.
     $this->drupalGet('admin/structure/taxonomy/manage/event_taxonomy/add');
+    $this->assertSession()->statusCodeEquals(200);
+
+    $this->drupalLogout();
+  }
+
+  /**
+   * Test whether the ecms_promotions feature installed properly.
+   */
+  private function ensurePromotionsFeatureInstalled(): void {
+    $account = $this->drupalCreateUser([
+      'create promotions content',
+    ]);
+    $this->drupalLogin($account);
+
+    // Ensure the promotions entity add form is available.
+    $this->drupalGet('node/add/promotions');
     $this->assertSession()->statusCodeEquals(200);
 
     $this->drupalLogout();